+++ lib/salah/location.rb
@@ -1,0 +1,4 @@
+class Salah::Location < Struct.new(:city, :country, :country_code,
+                                   :latitude, :longitude, :elevation,
+                                   :time_zone, :local_offset)
+end